requirements.txt
PyGithub>=2.0
requests>=2.30